+++ changelog.md
@@ -4,6 +4,7 @@
 - Catch bind address errors of xml rpc server
 - Refactor get_events, get_new_entities
 - Remove unnecessary checks
+- Separate enable/disable sysvar and program scan
 - Use paramset_description from channel
 
 # Version 2024.9.9 (2024-09-21)